requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==10.4.0
streamlit==1.38.0
requests==2.32.3
python-dotenv==1.0.1
aiofiles==23.2.1
